--- TESTINFO
@@ -1,1 +1,0 @@
-ExampleSet_manual depends on a real keyring and is excluded from automated tests.
